% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_make_grid.R
\name{smn_int_make_grid}
\alias{smn_int_make_grid}
\title{Build a local square grid around a point (km-based) and clip by radius
(robust CRS handling)}
\usage{
smn_int_make_grid(
  lat,
  lon,
  cell_km = 5,
  radius_km = 20,
  clip = c("intersect", "centroid"),
  crs_out = 4326,
  return = c("polygons", "centroids", "both")
)
}
\description{
Build a local square grid around a point (km-based) and clip by radius
(robust CRS handling)
}
